python/simple-flask-template/requirements.txt
coverage
cryptography
flake8
Flask-Cors
Flask-FeatureFlags
flask-marshmallow
Flask-RESTful
Flask-SQLAlchemy
flask-swagger
GeoAlchemy2
healthcheck
marshmallow-sqlalchemy
mutmut
psycopg2-binary
pytest
pytest-cov
python-json-logger
requests
Shapely
